feat #7 : 필터링 로직 useFilter에 정리 - 사용유무는 불확실
