+++ src/router/Router.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Landing from '../pages/Landing';
 import Reservation from '../pages/Reservation';
-import Test from "../pages/Test"
 
 type Props = {};
 
@@ -12,7 +11,6 @@ const Router = (props: Props) => {
       <Routes>
         <Route path='/' element={<Landing />} />
         <Route path='/reservation' element={<Reservation />} />
-        <Route path='test' element={<Test />} />
       </Routes>
     </BrowserRouter>
   );
