+++ src/types/hotelDataType.ts
@@ -5,10 +5,6 @@ export type BasicHotelDataType = {
     base: number;
     max: number;
   };
-  reservationDetail?: {
-    checkIn: string;
-    checkOut: string;
-  };
 };
 
 export type GetDataResultType = BasicHotelDataType[] | undefined;
